+++ tests/test_logging.py
@@ -1,4 +1,6 @@
-from research_town.utils.logging_callback import logging_callback
+from research_town.agents.agent_base import BaseResearchAgent
+from research_town.utils.logging import logging_callback
+from tests.constants import agent_profile_A
 
 
 def test_logging_callback() -> None:
